Swift 1.2 compatibility updates

+++ XCGLogger/Library/XCGLoggerTests/XCGLoggerTests.swift
@@ -166,7 +166,7 @@ class XCGLoggerTests: XCTestCase {
 
         let linesToLog = ["One", "Two", "Three", "Four", "Five", "Six", "Seven", "Eight", "Nine", "Ten"]
         let myConcurrentQueue = dispatch_queue_create("com.cerebralgardens.xcglogger.testMultiThreaded.queue", DISPATCH_QUEUE_CONCURRENT)
-        dispatch_apply(UInt(linesToLog.count), myConcurrentQueue) { (index: UInt) in
+        dispatch_apply(linesToLog.count, myConcurrentQueue) { (index: Int) in
             log.debug(linesToLog[Int(index)])
         }
     }
